Move Text loader to behat image-loaders configuration
